\name{mkcsv}
\alias{mkcsv}
\title{making csv file encoding = EUC-KR }
\usage{
mkcsv()
}
\description{
making csv file, encoding = EUC-KR
}

\examples{

  mkcsv(데이터셋, 'file명.csv' )  
        csv파일 저장 디폴트: EUC-KR  
# mkcsv(데이터셋, 'file명.csv' encoding = 'CP949')  
# mkcsv(데이터셋, 'file명.csv' encoding = 'UTF-8')  
  
}
